test(async-profiler): split GraphQL integration tests into separate file
